Refactor MBR validation and partition device creation methods for improved clarity and consistency
